Refactor company entity
